requirements.txt
asgiref==3.7.1
black==23.3.0
click==8.1.3
Django==4.2.1
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
Faker==18.9.0
geographiclib==2.0
geopy==2.3.0
Markdown==3.4.3
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg==3.1.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
